Working on movie detail screen

+++ app/src/main/java/com/example/movieapp/domain/model/Rating.kt
@@ -6,7 +6,7 @@ import kotlinx.parcelize.Parcelize
 @Parcelize
 data class Rating(
     val await: Double,
-    val filmCritics: Int,
+    val filmCritics: Double,
     val imdb: Double,
     val kp: Double,
     val russianFilmCritics: Double,
